casting_app/requirements.txt
# 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0
grad-cam>=1.4.8

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM - Claude
anthropic>=0.18.0
langchain>=0.1.0

# UI
streamlit>=1.28.0

# PDF 생성
reportlab>=4.0.0

# 시각화 (인터랙티브 차트)
plotly>=5.0.0

# 유틸리티
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0

# 실험 추적 (선택)
tensorboard>=2.14.0
wandb>=0.16.0
